Corrected notes now used in chuck with times

+++ music-space/Assets/Editor/NotesToTimesTest.cs
@@ -13,20 +13,20 @@ public class NewTestScript
         GameObject paintTest = new GameObject();
         Painter p = paintTest.AddComponent<Painter>();
 
-        Tuple<List<Double>, List<int>> result = p.NotesToTimes(new double[]{44,44,55,55,66});
+        Tuple<List<Double>, List<long>> result = p.NotesToTimes(new double[]{44,44,55,55,66});
         Assert.AreEqual(new double[]{44,55,66},result.Item1);
-        Assert.AreEqual(new int[]{200,200,100},result.Item2);
+        Assert.AreEqual(new long[]{200,200,100},result.Item2);
 
         result = p.NotesToTimes(new double[]{44,55,77,66});
         Assert.AreEqual(new double[]{44,55,77,66},result.Item1);
-        Assert.AreEqual(new int[]{100,100,100,100},result.Item2);
+        Assert.AreEqual(new long[]{100,100,100,100},result.Item2);
 
         result = p.NotesToTimes(new double[]{44,55,55,55,55,55,44});
         Assert.AreEqual(new double[]{44,55,44},result.Item1);
-        Assert.AreEqual(new int[]{100,500,100},result.Item2);
+        Assert.AreEqual(new long[]{100,500,100},result.Item2);
         
         result = p.NotesToTimes(new double[]{44,44,77,77});
         Assert.AreEqual(new double[]{44,77},result.Item1);
-        Assert.AreEqual(new int[]{200,200},result.Item2);
+        Assert.AreEqual(new long[]{200,200},result.Item2);
     }
 }
